// e-ROZ-front/src/app/models/hero.ts
export class Hero {
    id? : number;
    name : string;
    ability : string;
    description : string;
    status: any;
    strength: number;
    balance: number;
    deathInsurrane: number;
    cost: number;
    //add the other propreties after ability, description, strength, balance etc...
}